refactor: consolidate token configuration and improve token balance handling
